requirements-gpu.txt
chainer==6.2.0
cupy-cuda100==6.2.0
gin-config==0.1.4
disentanglement-lib==1.2
matplotlib==3.0.3
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
